frontend: store/company_forms: Rename {,_}partners.js to avoid name conflict
The module company_form/partner.js is acessible through
company_form.partner, which conflicts with the state
company_form.partner
