+++ frontend/store/company_forms/_partners.js
